Add files via upload
The Hackathon is scheduled for two days. Each team can spend 30 mins with the mentor on both days to sound off ideas/approaches. Mentor will NOT give away solutions. Each team need to select one problem statement :


Predict if the client will subscribe to direct
marketing campaign for a banking institution
Problem Statement:
The data is related to direct marketing campaigns of a Portuguese banking institution. Predict if
the client will subscribe to a term deposit based on a marketing campaign
Data Set Download:
https://drive.google.com/drive/folders/1urwTQPkUypJ6dGDJgS9Gszb83bfXEG6z?usp=sharing
Data Set Information:
The data is related to direct marketing campaigns of a Portuguese banking institution. The
marketing campaigns were based on phone calls. Often, more than one contact to the same
client was required, in order to access if the product (bank term deposit) would be ('yes') or not
('no') subscribed.
There are four datasets:
1. bank-additional-full.csv with all examples (41188) and 20 inputs, ordered by date (from
May 2008 to November 2010), very close to the data analyzed in [Moro et al., 2014]
2. bank-additional.csv with 10% of the examples (4119), randomly selected from 1), and 20
inputs.
3. bank-full.csv with all examples and 17 inputs, ordered by date (older version of this
dataset with fewer inputs).
4. bank.csv with 10% of the examples and 17 inputs, randomly selected from 3 (older
version of this dataset with fewer inputs). The smallest datasets are provided to test
more computationally demanding machine learning algorithms
Goal:- The classification goal is to predict if the client will subscribe (yes/no) a term deposit
(variable y).